backend/requirements.txt
backend==0.2.4.1
colorama==0.4.6
fastapi==0.110.1
llama_index==0.10.28
openai==1.17.0
psycopg2_binary==2.9.9
pydantic==2.7.0
pymilvus==2.4.0
pymongo==4.6.3
python-dotenv==1.0.1
python_bcrypt==0.3.2
striprtf==0.0.26
tabulate==0.9.0
uvicorn==0.29.0
